yolo-recognition/requirements.txt
absl-py==1.2.0
aiofiles==22.1.0
anyio==3.6.1
appnope==0.1.3
argcomplete==2.0.0
asttokens==2.0.8
astunparse==1.6.3
awscli==1.25.89
backcall==0.2.0
backports.cached-property==1.0.2
boto3==1.24.88
botocore==1.27.88
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
dacite==1.6.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.19
entrypoints==0.4
eventlet==0.33.1
executing==1.1.0
fiftyone==0.17.2
fiftyone-brain==0.9.1
fiftyone-db==0.3.0
flatbuffers==22.9.24
fonttools==4.37.4
future==0.18.2
gast==0.4.0
glob2==0.7
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphql-core==3.2.3
greenlet==1.1.3
grpcio==1.49.1
h11==0.12.0
h2==4.1.0
h5py==3.7.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.22.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kaleido==0.2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mongoengine==0.24.2
motor==3.0.0
ndjson==0.3.1
nest-asyncio==1.5.6
networkx==2.8.7
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
patool==1.12
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
pprintpp==0.4.0
priority==2.0.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
sentinel==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sse-starlette==1.1.6
sseclient-py==1.7.2
stack-data==0.5.1
starlette==0.21.0
strawberry-graphql==0.133.5
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.12
voxel51-eta==0.8.1
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
